dev-requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile dev-requirements.in
#
astroid==2.12.11
    # via pylint
attrs==21.4.0
    # via
    #   glom
    #   jsonschema
    #   semgrep
black==22.10.0
    # via -r dev-requirements.in
boltons==21.0.0
    # via
    #   face
    #   glom
    #   semgrep
bracex==2.3.post1
    # via wcmatch
build==0.8.0
    # via pip-tools
cachetools==5.2.0
    # via tox
certifi==2022.9.24
    # via requests
cfgv==3.3.1
    # via pre-commit
chardet==5.0.0
    # via tox
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   black
    #   click-option-group
    #   pip-tools
    #   semgrep
click-option-group==0.5.5
    # via semgrep
colorama==0.4.5
    # via
    #   build
    #   click
    #   pylint
    #   semgrep
    #   tox
    #   tqdm
defusedxml==0.7.1
    # via semgrep
dill==0.3.5.1
    # via pylint
distlib==0.3.6
    # via virtualenv
face==22.0.0
    # via glom
filelock==3.8.0
    # via virtualenv
flake8==5.0.4
    # via
    #   -r dev-requirements.in
    #   flake8-docstrings
flake8-docstrings==1.6.0
    # via -r dev-requirements.in
glom==22.1.0
    # via semgrep
identify==2.5.6
    # via pre-commit
idna==3.4
    # via requests
isort==5.10.1
    # via pylint
jsonschema==4.16.0
    # via semgrep
lazy-object-proxy==1.7.1
    # via astroid
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mypy==0.982
    # via -r dev-requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
nodeenv==1.7.0
    # via pre-commit
packaging==21.3
    # via
    #   build
    #   pyproject-api
    #   semgrep
    #   tox
pandas-stubs==1.5.0.221012
    # via -r dev-requirements.in
pathspec==0.10.1
    # via black
peewee==3.15.3
    # via semgrep
pep517==0.13.0
    # via build
pip-tools==6.9.0
    # via -r dev-requirements.in
platformdirs==2.5.2
    # via
    #   black
    #   pylint
    #   tox
    #   virtualenv
pluggy==1.0.0
    # via tox
pre-commit==2.20.0
    # via -r dev-requirements.in
pycodestyle==2.9.1
    # via flake8
pydocstyle==6.1.1
    # via flake8-docstrings
pyflakes==2.5.0
    # via flake8
pylint==2.15.4
    # via -r dev-requirements.in
pyparsing==3.0.9
    # via packaging
pyproject-api==1.1.1
    # via tox
pyrsistent==0.18.1
    # via jsonschema
python-lsp-jsonrpc==1.0.0
    # via semgrep
pyyaml==6.0
    # via pre-commit
requests==2.28.1
    # via semgrep
ruamel-yaml==0.17.21
    # via semgrep
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
semgrep==0.117.0
    # via -r dev-requirements.in
snowballstemmer==2.2.0
    # via pydocstyle
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via
    #   black
    #   build
    #   mypy
    #   pep517
    #   pylint
    #   pyproject-api
    #   semgrep
    #   tox
tomlkit==0.11.5
    # via pylint
tox==4.0.0b2
    # via -r dev-requirements.in
tqdm==4.64.1
    # via semgrep
types-pytz==2022.4.0.0
    # via pandas-stubs
typing-extensions==4.4.0
    # via
    #   mypy
    #   semgrep
ujson==5.5.0
    # via python-lsp-jsonrpc
urllib3==1.26.12
    # via
    #   requests
    #   semgrep
virtualenv==20.16.5
    # via
    #   pre-commit
    #   tox
wcmatch==8.4.1
    # via semgrep
wheel==0.37.1
    # via pip-tools
wrapt==1.14.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
